requirements.txt
crewai==0.28.8
fastapi==0.109.0
uvicorn==0.27.0
langchain-openai==0.0.5
python-dotenv==1.0.0
pydantic==2.5.0

